+++ server/controllers/recipes/index.js
@@ -1,4 +1,4 @@
 export {default as getUOM} from './get-uom.js';
 export {default as getById} from './get-by-id.js';
-export {default as getRecipes} from './get-recipes.js';
-export {default as saveRecipe} from './save-recipe.js';
+export {default as get} from './get.js';
+export {default as save} from './save.js';
