+++ server/lib/recipes/get.js
@@ -1,6 +1,6 @@
 import {Recipe} from '../../database-initi.js';
 
-export default async function getRecipes({memberId}) {
+export default async function get({memberId}) {
   return Recipe.findAndCountAll({
     where: {memberId}
   })
