+++ server/lib/recipes/index.js
@@ -1,4 +1,4 @@
 export {default as getUOMSvc} from './get-uom.js';
-export {default as getRecipesSvc} from './get-recipes.js';
+export {default as getSvc} from './get.js';
 export {default as getByIdSvc} from './get-by-id.js';
-export {default as saveRecipeSvc} from './save-recipe.js';
+export {default as saveSvc} from './set.js';
