+++ server/lib/recipes/set.js
@@ -1,5 +1,5 @@
 import {Recipe} from '../../database-initi.js';
 
-export default async function saveRecipe(recipe) {
+export default async function save(recipe) {
   return Recipe.create(recipe)
 }
