+++ server/routes/v1/recipes/index.js
@@ -1,11 +1,11 @@
 import express from 'express';
-import {getUOM, getById, getRecipes, saveRecipe} from '../../../controllers/recipes/index.js';
+import {getUOM, getById, get, save} from '../../../controllers/recipes/index.js';
 
 const router = new express.Router();
 
-router.post('/', saveRecipe);
+router.post('/', save);
 router.get('/ingredients/uom', getUOM);
 router.get('/:recipeId', getById);
-router.get('/', getRecipes);
+router.get('/', get);
 
 export default router;
